manage_telegrambot/requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==0.7
Django==4.2.5
django-ckeditor==6.7.0
django-crispy-forms==2.0
django-excel-response==2.0.5
django-flower==1.0.0
django-jinja==2.11.0
django-js-asset==2.1.0
et-xmlfile==1.1.0
flower==2.0.1
frozenlist==1.4.0
gevent==23.9.1
greenlet==3.0.0
humanize==0.5.1
idna==3.4
jdcal==1.4.1
Jinja2==3.1.2
kombu==5.3.2
loguru==0.7.2
magic-filter==1.0.12
MarkupSafe==2.1.3
multidict==6.0.4
openpyxl==2.4.11
Pillow==10.0.1
plumbum==1.8.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
# psycopg2==2.9.8
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rpyc==5.3.1
six==1.16.0
sqlparse==0.4.4
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
yarl==1.9.2
zope.event==5.0
zope.interface==6.1
